2. Apple Silicon (MPS) support for training, evaluation and quantization

+++ tinyml-tinyverse/tinyml_tinyverse/references/timeseries_regression/test_onnx.py
@@ -101,12 +101,12 @@ def main(gpu, args):
     logger.info(f"Loading ONNX model: {args.model_path}")
     ort_sess, input_name, output_name = load_onnx_model(args.model_path, args.generic_model)
 
-    predicted = torch.tensor([]).to(device, non_blocking=True)
-    ground_truth = torch.tensor([]).to(device, non_blocking=True)
+    predicted = torch.tensor([], dtype=torch.float32).to(device, non_blocking=True)
+    ground_truth = torch.tensor([], dtype=torch.float32).to(device, non_blocking=True)
 
     for _, batched_data, batched_target in data_loader:
-        batched_data = batched_data.to(device, non_blocking=True).float()
-        batched_target = batched_target.to(device, non_blocking=True).float()
+        batched_data = batched_data.float().to(device, non_blocking=True)
+        batched_target = batched_target.float().to(device, non_blocking=True)
         if transform:
             batched_data = transform(batched_data)
         for data in batched_data:
